change value to float for atmo_offset_factory[5];

+++ wolksensor/SDK/core/config.c
@@ -35,7 +35,7 @@ char mqtt_username[MQTT_USERNAME_SIZE];
 char mqtt_password[MQTT_PASSWORD_SIZE];
 
 float atmo_offset[3];
-uint32_t atmo_offset_factory[5];
+float atmo_offset_factory[5];
 
 bool load_device_id(void)
 {
@@ -323,7 +323,7 @@ bool load_offset_factory_status(void)
 {
 	if(global_dependencies.config_read(&atmo_offset_factory, CFG_OFFSET_FACTORY, 1, sizeof(atmo_offset_factory)))
 	{
-		LOG_PRINT(1, PSTR("Temperature offset factory is written: %d \n\rPressure offset factory is written: %d \n\rHumidity offset factory is written: %d \n\r"), atmo_offset_factory[0], atmo_offset_factory[1], atmo_offset_factory[2]);
+		LOG_PRINT(1, PSTR("Temperature offset factory is written: %0.2f \n\rPressure offset factory is written: %0.2f \n\rHumidity offset factory is written: %0.2f \n\r"), atmo_offset_factory[0], atmo_offset_factory[1], atmo_offset_factory[2]);
 		return true;
 	}
